introduce Database::g_async to reintroduce uv_ref()/uv_unref()
As mentioned in d150ffd9 in order to keep node alive while the database
connections were open I had to introduce a uv_async_t g_async static
member to the Database class. This is used by uv_async_init(), uv_ref()
and uv_unref() in versions of node > 0.7.9. In order to call
uv_async_init(), I had to add a callback watcher function called
WatcherCallback which is a static method of the Database class.