Fixes for #30; Attempt number two: Decimals and VarChars.
*Add `decimals` member to the Parameter struct and use it for SQLBindParameters
*Don't use SQL_UNKNOWN_TYPE for Null parmeters, instead use SQL_VARCHAR

This should fix #30, however the values I have chosen for `.decimals` and `.size`
need to be addressed for the `IsNumber()` Parameter case. We may have to do
something like count the digits to the left of the decimal point to get `.size`
and count the digits to the right to get `.decimals`.