Only free ODBCStatement parameters on Free(), Bind() and BindSync()
If parameters have been previously been bound, then the should be freed
before any new parameters are created. This would happen in Bind() or
BindSync(). They should also be freed in the dtor of the ODBCStatment
class.

This will allow us to bind once and repeatedly execte() and insert the
same values in case anyone needs to do this for some reason.