test: replace open() and close() calls with openSync() and closeSync()
Only made the replacement in places where we are not explicitly
testing the open() or close() functionality. The reason behind
this is that MySQL has a bug which causes most of the tests to
time out because a thread is not released if a connection is
opened and closed from two different threads. This tends to occur
when we are using the async calls because the ODBC open/close call
occurs on the thread pool. And we can't guarantee that it will
be on the same thread. The only way to guarantee ODBC open/close
is called from the same thread is if we call it from the main
thread. In order to do that, we must use the Sync versions of
open() or close()