Downgraded tar.gz module to 1.0.2
The ibm_db tests fail on linux for z systems as well as sometimes on aix ppc. We have discovered this is due to a bug introduced in the tar.gz module from version 1.0.3 that has not been fixed yet. So we suggest keeping the module tar.gz@1.0.2 until it works properly on future versions.
Another solution is to use the targz module instead of tar.gz as that functions properly.