Add z/OS specific updates
A set of changes for z/OS ODBC compatiability:
- Update default szTableType type to upper-case, as z/OS treats
them as keywords, and requires them to be capitalized.
- Update timeInfo struct initization to match xlC spec.
- Added size 0 check for malloc, as xlC behaviour in that case
is to return NULL.
- Change diagnostics record checking to loop until SQL_NO_DATA is found.
ODBC driver on z/OS does not support SQLGetDiagField.
- For catalog APIs, the catalog qualifier (szTableQualifier) and
its length (cbTableQualifier) must be set to NULL and 0, respectively.
Three part naming for catalog APIs are not supported on z/OS.
- Update printf diagnostic to use the _a versions on z/OS, as String
literals are treated as ASCII when compiling this native add-on.

Signed-off-by: Joran Siu <joransiu@ca.ibm.com>