Restrict createDbSync/dropDbSync on z/OS
The createDbSync and dropDbSync APIs are not applicable to z/OS.
Update the API to throw an error on z/OS if the APIs are invoked.

The reason that they are not supported is because there is a
fundamental difference in how a database is implemented/used
in DB2 for z/OS and DB2 for LUW.  A database in z/OS is simply
a logical collection of table/index spaces that you create
using the "CREATE DATABASE" SQL statement.  A database in LUW
is conceptually equivalent to a subsystem in z/OS.  When working
with DB2 for LUW, you connect to a specific database, whereas
in DB2 for z/OS, you connect to a DB2 subsystem.  So, if the
purpose of using SQLCreateDb is to create a database, so that
you can connect to it later, creating a database on z/OS will
not serve that purpose.

Signed-off-by: Joran Siu <joransiu@ca.ibm.com>